Grosse mise à jour

+++ views/listCommande.php
@@ -1,16 +1,22 @@
-<a href="./index.php?action=formAddCommande">Ajouter</a>
+<?php if($_SESSION['grade'] == 1){ ?>
+  <a href="./index.php?action=formAddCommande">Ajouter</a>
+<?php } ?>
 <table>
   <thead>
     <th>Id</th>
     <th>Référence</th>
     <th>Civilité</th>
-    <th>Nom Client</th>
-    <th>Prénom Client</th>
+    <th>Nom</th>
+    <th>Prénom</th>
+    <th>Produit</th>
+    <th>Quantite</th>
     <th>Date Commande</th>
     <th>Date Expédition</th>
-    <th>Statut Commande</th>
-    <th>Editer</th>
-    <th>Supprimer</th>
+    <th>Statut</th>
+    <?php if($_SESSION['grade'] == 1){ ?>
+      <th>Editer</th>
+      <th>Supprimer</th>
+    <?php } ?>
   </thead>
   <tbody>
     <?php
@@ -21,11 +27,15 @@
           echo '<td>' . $commande->getClient()->getCivilite() . '</td>';
           echo '<td>' . $commande->getClient()->getNom() . '</td>';
           echo '<td>' . $commande->getClient()->getPrenom() . '</td>';
+          echo '<td>' . $commande->getProduit()->getLibelle() . '</td>';
+          echo '<td>' . $commande->getCommandeProduit()->getQuantite() . '</td>';
           echo '<td>' . $commande->getDateCommande() . '</td>';
           echo '<td>' . $commande->getDateExpedition() . '</td>';
           echo '<td>' . $commande->getStatut()->getLibelle() . '</td>';
-          echo '<td><a href="./index.php?action=formEditCommande&id=' . $commande->getId() . '"">Editer</a></td>';
-          echo '<td><a href="./index.php?action=deleteCommande&id=' . $commande->getId() . '">Supprimer</a></td>';
+          if($_SESSION['grade'] == 1){
+            echo '<td><a href="./index.php?action=formEditCommande&id=' . $commande->getId() . '"">Editer</a></td>';
+            echo '<td><a href="./index.php?action=deleteCommande&id=' . $commande->getId() . '">Supprimer</a></td>';
+          }
         echo '</tr>';
       }
     ?>
