+++ views/listProduit.php
@@ -1,14 +1,18 @@
-<a href="./index.php?action=formAddCommande">Ajouter</a>
+<?php if($_SESSION['grade'] == 1){ ?>
+  <a href="./index.php?action=formAddCommande">Ajouter</a>
+<?php } ?>
 <table>
   <thead>
     <th>Id</th>
     <th>Référence</th>
     <th>Libellé</th>
-    <th>Description</th>   
+    <th>Description</th>
     <th>Prix Unitaire</th>
     <th>Quantité en stock</th>
-    <th>Editer</th>
-    <th>Supprimer</th>
+    <?php if($_SESSION['grade'] == 1){ ?>
+      <th>Editer</th>
+      <th>Supprimer</th>
+    <?php } ?>
   </thead>
   <tbody>
     <?php
@@ -20,8 +24,10 @@
           echo '<td>' . $produit->getDescription() . '</td>';
           echo '<td>' . $produit->getPrixUnitaire() . ' €</td>';
           echo '<td>' . $produit->getQuantite() . '</td>';
-          echo '<td><a href="./index.php?action=formEditCommande&id=' . $produit->getId() . '"">Editer</a></td>';
-          echo '<td><a href="./index.php?action=deleteCommande&id=' . $produit->getId() . '">Supprimer</a></td>';
+          if($_SESSION['grade'] == 1){
+            echo '<td><a href="./index.php?action=formEditCommande&id=' . $produit->getId() . '"">Editer</a></td>';
+            echo '<td><a href="./index.php?action=deleteCommande&id=' . $produit->getId() . '">Supprimer</a></td>';
+          }
         echo '</tr>';
       }
     ?>
